// src/app/core/api/generated/model/patientSearchCriteriaDto.ts
/**
 * Patient Account Tool API
 *
 * Contact: platform@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PatientSearchCriteriaDto { 
    patientId?: number;
    mrn?: string;
    firstName?: string;
    lastName?: string;
    dateOfBirth?: string;
    email?: string;
    phone?: string;
    search?: string;
    status?: string;
    admissionStatus?: string;
}